+++ frontend/client/index.html
@@ -34,7 +34,7 @@
 <script src="/assets/dat.gui.min.js"></script>
 <script src="/assets/stats.min.js"></script>
 
-<script src="script.js"></script>
+<script type="module" src="script.js"></script>
 
 </body>
 </html>
