+++ frontend/client/script.js
@@ -1,7 +1,10 @@
 'use strict';
 
+import { testExport } from './animation.js';
+import { onDocumentKeyDown, onDocumentMouseUp, onDocumentMouseDown } from './keypress.js';
+import { stageGui, initGui } from './gui.js';
+
 let container, stats;
-let stageGui, connectionGui;
 let camera, scene, renderer;
 let topDirectionalLight, leftDirectionalLight, rightDirectionalLight;
 let mesh, lines, geometry;
@@ -393,17 +396,6 @@ let getControl = () => {
     return control;
 };
 
-let initGui = () => {
-    connectionGui = new dat.GUI( { width: 200 } );
-    stageGui = new dat.GUI( { width: 200 } );
-    stageGui.add({ AddLinearStage: () => {
-        addLinearStage();
-    } }, 'AddLinearStage');
-    stageGui.add({ AddRotaryStage: () => {
-        addRotaryStage();
-    } }, 'AddRotaryStage');
-};
-
 let _getIntersectsFromClickWithCandidates = (event, candidates) => {
     let vector = new THREE.Vector3();
     let raycaster = new THREE.Raycaster();
@@ -468,93 +460,6 @@ let getRootConnectionGroup = (group) => {
 /**
  * KEYPRESS LOGIC
  */
-let onDocumentMouseDown = (event) => {
-    // NOTE: do not fire click events if we click on the GUI
-    if (stageGui.domElement.contains(event.target)) {
-        if (event.target.className === "title") {
-            let maybeStage = findStageWithName(event.target.innerHTML);
-            if (maybeStage !== undefined) {
-                destroyControl();
-                generateControlForGroup(maybeStage);
-                focus(maybeStage);
-            }
-        }
-        return;
-    }
-
-    let candidates = getStages().concat(getTool()).concat(getConnectionHandles());
-    let isectGroups = _getIntersectsFromClickWithCandidates(event, candidates);
-    let isectControl;
-    if (getControl() === undefined) {
-        isectControl = [];
-    }
-    else {
-        isectControl = _getIntersectsFromClickWithCandidates(event, [getControl()]);
-    }
-    let possibleHandles = isectGroups.filter((result) => result.object.name === 'connectionHandle')
-    if (possibleHandles.length > 0) {
-        let currHandle = possibleHandles[0];
-        if (activeSelectionHandle === undefined) {
-            setActiveSelectionHandle(currHandle);
-        } else {
-            let fromModule;
-            if (activeSelectionHandle.object.parent.isTool) {
-                fromModule = getTool();
-            } else {
-                fromModule = findStageWithName(activeSelectionHandle.object.parent.stageName);
-            }
-            let fromPlace = activeSelectionHandle.object.place;
-            let toStage = findStageWithName(currHandle.object.parent.stageName);
-            let toPlace = currHandle.object.place;
-            connectParentChild(fromModule, fromPlace, toStage, toPlace);
-            releaseActiveSelectionHandle();
-        }
-    }
-    // Kludge: isectControl length >= 3 means we are clicking the controls
-    if (isectControl.length < 3 && isectGroups.length > 0) {
-        let stage = getObjectGroup(isectGroups[0].object);
-        // If we are holding shift, make a connection
-        if (event.shiftKey) {
-            if (getFocus().isTool) {
-                connectToolToStage(getFocus(), stage);
-            }
-            else {
-                let parentStageName = getStageName(getFocus());
-                let childStageName = getStageName(stage);
-                let place = prompt(`Where is ${parentStageName} connecting to ${childStageName}?`);
-                if (!(place === 'platform' || place === 'right' || place === 'left')) {
-                    return;
-                }
-                connectParentChild(getFocus(), stage, place);
-            }
-        }
-
-        // Otherwise, just focus the new stage
-        destroyControl();
-        generateControlForGroup(stage);
-        focus(stage);
-        openFolderForStage(stage);
-    }
-    else if (isectControl.length < 3) {
-        unfocus();
-        destroyControl();
-        openFolderForStage(null);
-    }
-};
-
-let onDocumentMouseUp = (event) => {
-    // If we had clicked on a stage folder, close all folders and let the
-    // datGui mouseup handler just open the one stage folder
-    if (stageGui.domElement.contains(event.target)) {
-        if (event.target.className === "title") {
-            openFolderForStage(null);
-        }
-        return;
-    }
-    // FIXME: better to do only on rotation, but this is easier
-    redetermineAllStageAxes();
-    redetermineAccepts();
-};
 
 let openFolderForStage = (stage) => {
     let groups = getStages();
@@ -568,29 +473,6 @@ let openFolderForStage = (stage) => {
             }
         }
     });
-};
-
-let onDocumentKeyDown = (event) => {
-    if (event.target.nodeName === "PRE" || event.target.nodeName === "INPUT") {
-        if (event.key === "Escape") {
-            document.activeElement.blur();
-        }
-        return;
-    }
-    if (event.key === "Backspace") {
-        if (getFocus() !== null && event.shiftKey) {
-            deleteStage(getFocus());
-        }
-    }
-    if (event.key === "m") {
-        swapControlMode();
-    }
-    if (event.key === "s") {
-        toggleConnectionHandles();
-    }
-    if (event.key === "Escape") {
-        releaseActiveSelectionHandle();
-    }
 };
 
 let initCamera = () => {
@@ -607,8 +489,8 @@ let initCamera = () => {
 };
 
 let initScene = () => {
+    console.log(testExport);
     scene = new THREE.Scene();
-    clock = new THREE.Clock();
     scene.background = new THREE.Color(0xf5f6f8);
     topDirectionalLight = new THREE.DirectionalLight( 0xffffff, 1.00 );
     leftDirectionalLight = new THREE.DirectionalLight( 0xffffff, 0.75 );
